Fixing gamedev-net/nehe-opengl#3.

+++ linuxglx/lesson07/lesson07.c
@@ -76,10 +76,10 @@ int loadBMP(char *filename, textureImage *texture)
 {
     FILE *file;
     unsigned short int bfType;
-    long int bfOffBits;
+    int bfOffBits;
     short int biPlanes;
     short int biBitCount;
-    long int biSizeImage;
+    int biSizeImage;
     int i;
     unsigned char temp;
     /* make sure the file is there and open it read-only (binary) */
@@ -103,12 +103,12 @@ int loadBMP(char *filename, textureImage *texture)
     /* skip file size and reserved fields of bitmap file header */
     fseek(file, 8, SEEK_CUR);
     /* get the position of the actual bitmap data */
-    if (!fread(&bfOffBits, sizeof(long int), 1, file))
+    if (!fread(&bfOffBits, sizeof(int), 1, file))
     {
         printf("Error reading file!\n");
         return 0;
     }
-    printf("Data at Offset: %ld\n", bfOffBits);
+    printf("Data at Offset: %d\n", bfOffBits);
     /* skip size of bitmap info header */
     fseek(file, 4, SEEK_CUR);
     /* get the width of the bitmap */
@@ -138,7 +138,7 @@ int loadBMP(char *filename, textureImage *texture)
     }
     /* calculate the size of the image in bytes */
     biSizeImage = texture->width * texture->height * 3;
-    printf("Size of the image data: %ld\n", biSizeImage);
+    printf("Size of the image data: %d\n", biSizeImage);
     texture->data = malloc(biSizeImage);
     /* seek to the actual data */
     fseek(file, bfOffBits, SEEK_SET);
